broke api callbacks into seperate folder, fixed getTwitterFeed action
