+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/Team-Tenacity/domonation#readme",
   "dependencies": {
+    "axios": "^0.13.1",
     "babel-core": "^6.10.4",
     "babel-loader": "^6.2.4",
     "babel-polyfill": "^6.9.1",
